requirements.txt
flask
tensorflow
pillow
matplotlib
seaborn
pandas
keras-tuner